+++ action-packed-react/package.json
@@ -3,7 +3,8 @@
   "version": "0.0.0",
   "description": "",
   "keywords": [],
-  "main": "./disk/lib/index.js",
+  "main": "./dist/lib/index.js",
+  "browser": "./dist/lib/index.js",
   "module": "./dist/lib/index.js",
   "types": "./dist/types/",
   "author": "Eric Wooley <ericwooley@gmail.com>",
