+++ create-apr/bin.js
@@ -14,7 +14,9 @@ try {
 const log = debug('apr:create')
 log('argv', argv._)
 process.env.HYGEN_TMPLS = join(__dirname, './template/_templates/')
-spawnSync(hygen, ['apr-init', 'new', '--name', argv.dev ? 'playground' : argv.name || 'my-app'], { stdio: 'inherit' })
+spawnSync(hygen, ['apr-init', 'new', '--name', argv.dev ? 'playground' : argv.name || 'my-app'], {
+  stdio: 'inherit'
+})
 
 if (argv.dev) {
   console.log('installing projects')
@@ -31,11 +33,10 @@ if (argv.dev) {
   spawnSync('yarn', ['link'], { stdio: 'inherit' })
   process.cwd(currentPath)
   spawnSync('yarn', ['add', aprCliPath, aprPath], { stdio: 'inherit' })
-  spawnSync('yarn', ['link', 'apr-cli'], { stdio: 'inherit' })
 } else {
   spawnSync('yarn', [], { stdio: 'inherit' })
-  // spawnSync('yarn', ['add', 'apr-cli'], { stdio: 'inherit' })
-  // spawnSync('yarn', ['add', 'action-packed-react'], {
-  // stdio: 'inherit'
-  // })
+  spawnSync('yarn', ['add', 'apr-cli'], { stdio: 'inherit' })
+  spawnSync('yarn', ['add', 'action-packed-react'], {
+    stdio: 'inherit'
+  })
 }
